add productModifierItem and productModifierItemSelection fragments
