+++ Core/GraphQL/API/sell_createMyBasket.graphql.swift
@@ -28,6 +28,8 @@ public extension ApolloType {
       document.append("\n" + FragmentTimeslot.fragmentDefinition)
       document.append("\n" + FragmentProductVariant.fragmentDefinition)
       document.append("\n" + FragmentProduct.fragmentDefinition)
+      document.append("\n" + FragmentProductModifierItemSelection.fragmentDefinition)
+      document.append("\n" + FragmentProductModifierItem.fragmentDefinition)
       return document
     }
