+++ Core/GraphQL/API/sell_fragment_basket.graphql.swift
@@ -59,21 +59,7 @@ public extension ApolloType {
           }
           productModifierItems {
             __typename
-            productModifierItem {
-              __typename
-              id
-              status
-              externalId
-              additionalPrice
-              position
-              translations {
-                __typename
-                language
-                title
-              }
-            }
-            quantity
-            totalPrice
+            ...FragmentProductModifierItemSelection
           }
         }
       }
@@ -683,9 +669,7 @@ public extension ApolloType {
         public static var selections: [GraphQLSelection] {
           return [
             GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-            GraphQLField("productModifierItem", type: .object(ProductModifierItem.selections)),
-            GraphQLField("quantity", type: .scalar(Int.self)),
-            GraphQLField("totalPrice", type: .scalar(Int.self)),
+            GraphQLFragmentSpread(FragmentProductModifierItemSelection.self),
           ]
         }
 
@@ -693,10 +677,6 @@ public extension ApolloType {
 
         public init(unsafeResultMap: ResultMap) {
           self.resultMap = unsafeResultMap
-        }
-
-        public init(productModifierItem: ProductModifierItem? = nil, quantity: Int? = nil, totalPrice: Int? = nil) {
-          self.init(unsafeResultMap: ["__typename": "ProductModifierItemSelection", "productModifierItem": productModifierItem.flatMap { (value: ProductModifierItem) -> ResultMap in value.resultMap }, "quantity": quantity, "totalPrice": totalPrice])
         }
 
         public var __typename: String {
@@ -708,167 +688,28 @@ public extension ApolloType {
           }
         }
 
-        public var productModifierItem: ProductModifierItem? {
+        public var fragments: Fragments {
           get {
-            return (resultMap["productModifierItem"] as? ResultMap).flatMap { ProductModifierItem(unsafeResultMap: $0) }
+            return Fragments(unsafeResultMap: resultMap)
           }
           set {
-            resultMap.updateValue(newValue?.resultMap, forKey: "productModifierItem")
+            resultMap += newValue.resultMap
           }
         }
 
-        public var quantity: Int? {
-          get {
-            return resultMap["quantity"] as? Int
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "quantity")
-          }
-        }
-
-        public var totalPrice: Int? {
-          get {
-            return resultMap["totalPrice"] as? Int
-          }
-          set {
-            resultMap.updateValue(newValue, forKey: "totalPrice")
-          }
-        }
-
-        public struct ProductModifierItem: GraphQLSelectionSet {
-          public static let possibleTypes: [String] = ["ProductModifierItem"]
-
-          public static var selections: [GraphQLSelection] {
-            return [
-              GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-              GraphQLField("id", type: .nonNull(.scalar(GraphQLID.self))),
-              GraphQLField("status", type: .scalar(ProductModifierItemStatus.self)),
-              GraphQLField("externalId", type: .scalar(String.self)),
-              GraphQLField("additionalPrice", type: .scalar(Int.self)),
-              GraphQLField("position", type: .scalar(Int.self)),
-              GraphQLField("translations", type: .list(.object(Translation.selections))),
-            ]
-          }
-
+        public struct Fragments {
           public private(set) var resultMap: ResultMap
 
           public init(unsafeResultMap: ResultMap) {
             self.resultMap = unsafeResultMap
           }
 
-          public init(id: GraphQLID, status: ProductModifierItemStatus? = nil, externalId: String? = nil, additionalPrice: Int? = nil, position: Int? = nil, translations: [Translation?]? = nil) {
-            self.init(unsafeResultMap: ["__typename": "ProductModifierItem", "id": id, "status": status, "externalId": externalId, "additionalPrice": additionalPrice, "position": position, "translations": translations.flatMap { (value: [Translation?]) -> [ResultMap?] in value.map { (value: Translation?) -> ResultMap? in value.flatMap { (value: Translation) -> ResultMap in value.resultMap } } }])
-          }
-
-          public var __typename: String {
+          public var fragmentProductModifierItemSelection: FragmentProductModifierItemSelection {
             get {
-              return resultMap["__typename"]! as! String
+              return FragmentProductModifierItemSelection(unsafeResultMap: resultMap)
             }
             set {
-              resultMap.updateValue(newValue, forKey: "__typename")
-            }
-          }
-
-          public var id: GraphQLID {
-            get {
-              return resultMap["id"]! as! GraphQLID
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "id")
-            }
-          }
-
-          public var status: ProductModifierItemStatus? {
-            get {
-              return resultMap["status"] as? ProductModifierItemStatus
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "status")
-            }
-          }
-
-          public var externalId: String? {
-            get {
-              return resultMap["externalId"] as? String
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "externalId")
-            }
-          }
-
-          public var additionalPrice: Int? {
-            get {
-              return resultMap["additionalPrice"] as? Int
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "additionalPrice")
-            }
-          }
-
-          public var position: Int? {
-            get {
-              return resultMap["position"] as? Int
-            }
-            set {
-              resultMap.updateValue(newValue, forKey: "position")
-            }
-          }
-
-          public var translations: [Translation?]? {
-            get {
-              return (resultMap["translations"] as? [ResultMap?]).flatMap { (value: [ResultMap?]) -> [Translation?] in value.map { (value: ResultMap?) -> Translation? in value.flatMap { (value: ResultMap) -> Translation in Translation(unsafeResultMap: value) } } }
-            }
-            set {
-              resultMap.updateValue(newValue.flatMap { (value: [Translation?]) -> [ResultMap?] in value.map { (value: Translation?) -> ResultMap? in value.flatMap { (value: Translation) -> ResultMap in value.resultMap } } }, forKey: "translations")
-            }
-          }
-
-          public struct Translation: GraphQLSelectionSet {
-            public static let possibleTypes: [String] = ["ProductModifierItemTranslation"]
-
-            public static var selections: [GraphQLSelection] {
-              return [
-                GraphQLField("__typename", type: .nonNull(.scalar(String.self))),
-                GraphQLField("language", type: .scalar(Language.self)),
-                GraphQLField("title", type: .scalar(String.self)),
-              ]
-            }
-
-            public private(set) var resultMap: ResultMap
-
-            public init(unsafeResultMap: ResultMap) {
-              self.resultMap = unsafeResultMap
-            }
-
-            public init(language: Language? = nil, title: String? = nil) {
-              self.init(unsafeResultMap: ["__typename": "ProductModifierItemTranslation", "language": language, "title": title])
-            }
-
-            public var __typename: String {
-              get {
-                return resultMap["__typename"]! as! String
-              }
-              set {
-                resultMap.updateValue(newValue, forKey: "__typename")
-              }
-            }
-
-            public var language: Language? {
-              get {
-                return resultMap["language"] as? Language
-              }
-              set {
-                resultMap.updateValue(newValue, forKey: "language")
-              }
-            }
-
-            public var title: String? {
-              get {
-                return resultMap["title"] as? String
-              }
-              set {
-                resultMap.updateValue(newValue, forKey: "title")
-              }
+              resultMap += newValue.resultMap
             }
           }
         }
