+++ Core/GraphQL/API/sell_getOrders.graphql.swift
@@ -32,6 +32,8 @@ public extension ApolloType {
       document.append("\n" + FragmentFulfilmentPointCategory.fragmentDefinition)
       document.append("\n" + FragmentTimeslot.fragmentDefinition)
       document.append("\n" + FragmentProductVariant.fragmentDefinition)
+      document.append("\n" + FragmentProductModifierItemSelection.fragmentDefinition)
+      document.append("\n" + FragmentProductModifierItem.fragmentDefinition)
       return document
     }
