+++ Core/GraphQL/GraphQLFiles/sell_fragment_productModifierItem.graphql
@@ -1,0 +1,11 @@
+fragment FragmentProductModifierItem on ProductModifierItem {
+  id
+  status
+  externalId
+  additionalPrice
+  position
+  translations {
+    language
+    title
+  }
+}
