+++ Core/GraphQL/GraphQLFiles/sell_fragment_productModifierItemSelection.graphql
@@ -1,0 +1,5 @@
+fragment FragmentProductModifierItemSelection on ProductModifierItemSelection {
+  productModifierItem { ...FragmentProductModifierItem }
+  quantity
+  totalPrice
+}
